# docs/swagger.yaml
basePath: /api/v1        // Base path for all API routes
definitions:
...
  types.CreateUserPreferenceProfileParams:
    properties:
      budget_level:
        type: integer
      dietary_needs:
        items:
          type: string
        type: array
      interests:
        items:
          type: string
        type: array
      is_default:
        description: Default is FALSE in DB
        type: boolean
      prefer_accessible_pois:
        type: boolean
      prefer_dog_friendly:
        type: boolean
      prefer_outdoor_seating:
        type: boolean
      preferred_pace:
        $ref: '#/definitions/types.SearchPace'
      preferred_time:
        $ref: '#/definitions/types.DayPreference'
      preferred_transport:
        $ref: '#/definitions/types.TransportPreference'
      preferred_vibes:
        description: Use empty slice if not provided?
        items:
          type: string
        type: array
      profile_name:
        type: string
      search_radius_km:
        type: number
      tags:
        items:
          type: string
        type: array
      user_id:
        description: Added for clarity
        type: string
    required:
    - profile_name
    - user_id
    type: object
  types.DayPreference:
    enum:
    - any
    - day
    - night
    type: string
    x-enum-comments:
      DayPreferenceAny: No specific preference
      DayPreferenceDay: Primarily daytime activities
      DayPreferenceNight: Primarily evening/night activities
    x-enum-varnames:
    - DayPreferenceAny
    - DayPreferenceDay
    - DayPreferenceNight
  types.Interest:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      description:
        description: Use pointer if nullable
        type: string
      id:
        type: string
      name:
        type: string
      source:
        type: string
      updated_at:
        type: string
    type: object
...
  types.Response:
    properties:
      error:
        description: Optional error message.
        example: Resource not found
        type: string
      message:
        description: Optional success message.
        example: Operation successful
        type: string
      success:
        description: Indicates if the operation was successful.
        example: true
        type: boolean
    type: object
  types.SearchPace:
    enum:
    - any
    - relaxed
    - moderate
    - fast
    type: string
    x-enum-comments:
      SearchPaceAny: No preference
      SearchPaceFast: Pack in many activities
      SearchPaceModerate: Standard pace
      SearchPaceRelaxed: Fewer, longer activities
    x-enum-varnames:
    - SearchPaceAny
    - SearchPaceRelaxed
    - SearchPaceModerate
    - SearchPaceFast
  types.Tags:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      source:
        type: string
      tag_type:
        description: Consider using a specific enum type
        type: string
      updated_at:
        type: string
    type: object
...
  types.TransportPreference:
    enum:
    - any
    - walk
    - public
    - car
    type: string
    x-enum-varnames:
    - TransportPreferenceAny
    - TransportPreferenceWalk
    - TransportPreferencePublic
    - TransportPreferenceCar
...
  types.UpdateUserPreferenceProfileParams:
    properties:
      budget_level:
        type: integer
      dietary_needs:
        items:
          type: string
        type: array
      interests:
        items:
          type: string
        type: array
      is_default:
        description: Default is FALSE in DB
        type: boolean
      prefer_accessible_pois:
        type: boolean
      prefer_dog_friendly:
        type: boolean
      prefer_outdoor_seating:
        type: boolean
      preferred_pace:
        $ref: '#/definitions/types.SearchPace'
      preferred_time:
        $ref: '#/definitions/types.DayPreference'
      preferred_transport:
        $ref: '#/definitions/types.TransportPreference'
      preferred_vibes:
        description: Use empty slice if not provided?
        items:
          type: string
        type: array
      profile_name:
        type: string
      search_radius_km:
        type: number
      tags:
        items:
          type: string
        type: array
      updated_at:
        description: Optional, can be set to nil
        type: string
    required:
    - profile_name
    type: object
...
  types.UserPreferenceProfileResponse:
    properties:
      budget_level:
        type: integer
      created_at:
        type: string
      dietary_needs:
        description: Assuming TEXT[] maps to []string
        items:
          type: string
        type: array
      id:
        type: string
      interests:
        description: Interests linked to this profile
        items:
          $ref: '#/definitions/types.Interest'
        type: array
      is_default:
        type: boolean
      prefer_accessible_pois:
        type: boolean
      prefer_dog_friendly:
        type: boolean
      prefer_outdoor_seating:
        type: boolean
      preferred_pace:
        $ref: '#/definitions/types.SearchPace'
      preferred_time:
        $ref: '#/definitions/types.DayPreference'
      preferred_transport:
        $ref: '#/definitions/types.TransportPreference'
      preferred_vibes:
        description: Assuming TEXT[] maps to []string
        items:
          type: string
        type: array
      profile_name:
        type: string
      search_radius_km:
        type: number
      tags:
        description: Tags to avoid for this profile
        items:
          $ref: '#/definitions/types.Tags'
        type: array
      updated_at:
        type: string
      user_id:
        description: Might omit in some API responses
        type: string
      user_latitude:
        type: number
      user_longitude:
        type: number
    type: object
...
host: localhost:8000 // Adjust to your actual host/port
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for personalized city discovery and recommendations.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: WanderWiseAI API
  version: "1.0"
paths:
...
  /user/interests:
    get:
      consumes:
      - application/json
      description: Retrieves all available interests.
      produces:
      - application/json
      responses:
        "200":
          description: All Interests
          schema:
            items:
              $ref: '#/definitions/types.Interest'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.Response'
      summary: Get All Interests
      tags:
      - User
...
  /user/profiles:
    get:
      consumes:
      - application/json
      description: Fetches all preference profiles for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: User Preference Profiles
          schema:
            items:
              $ref: '#/definitions/types.UserPreferenceProfileResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: Get User Preference Profiles
      tags:
      - User Profiles
    post:
      consumes:
      - application/json
      description: Creates a new preference profile for the authenticated user
      parameters:
      - description: Profile Creation Parameters
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/types.CreateUserPreferenceProfileParams'
      produces:
      - application/json
      responses:
        "201":
          description: Created Profile
          schema:
            $ref: '#/definitions/types.UserPreferenceProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: Create User Preference Profile
      tags:
      - User Profiles
  /user/profiles/{profileID}:
    delete:
      consumes:
      - application/json
      description: Deletes an existing preference profile for the authenticated user
      parameters:
      - description: Profile ID
        in: path
        name: profileID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: Delete User Preference Profile
      tags:
      - User Profiles
    get:
      consumes:
      - application/json
      description: Fetches a specific preference profile for the authenticated user
      parameters:
      - description: Profile ID
        in: path
        name: profileID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User Preference Profile
          schema:
            $ref: '#/definitions/types.UserPreferenceProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: Get User Preference Profile
      tags:
      - User Profiles
    put:
      consumes:
      - application/json
      description: Updates an existing preference profile for the authenticated user
      parameters:
      - description: Profile ID
        in: path
        name: profileID
        required: true
        type: string
      - description: Profile Update Parameters
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/types.UpdateUserPreferenceProfileParams'
      produces:
      - application/json
      responses:
        "200":
          description: Updated Profile
          schema:
            $ref: '#/definitions/types.UserPreferenceProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: Update User Preference Profile
      tags:
      - User Profiles
  /user/profiles/default:
    get:
      consumes:
      - application/json
      description: Fetches the default preference profile for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Default User Preference Profile
          schema:
            $ref: '#/definitions/types.UserPreferenceProfileResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: Get Default User Preference Profile
      tags:
      - User Profiles
  /user/profiles/default/{profileID}:
    put:
      consumes:
      - application/json
      description: Sets an existing preference profile as the default for the authenticated
        user
      parameters:
      - description: Profile ID
        in: path
        name: profileID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Default User Preference Profile
          schema:
            $ref: '#/definitions/types.UserPreferenceProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: Set Default User Preference Profile
      tags:
      - User Profiles
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"